ethernet in seperate phys sub-type
